Rename module & cleanup
