+++ src/main/java/edu/kit/datamanager/pit/pidsystem/impl/local/PidDatabaseObjectDao.java
@@ -1,11 +1,9 @@
-package edu.kit.datamanager.pit.pidsystem.impl.sandbox;
+package edu.kit.datamanager.pit.pidsystem.impl.local;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-import edu.kit.datamanager.pit.pidlog.KnownPid;
 
 /**
  * Object to access PID records from the database.
